move project apis out of auth folder
