+++ pages/api/projects/[id]/index.js
@@ -1,7 +1,7 @@
-import dbConnect from "../../../utils/dbConnect";
-import Project from "../../../models/Project.js";
+import dbConnect from "../../../../utils/dbConnect";
+import Project from "../../../../models/Project.js";
 import { getServerSession } from "next-auth/next";
-import { authOptions } from "../auth/[...nextauth]";
+import { authOptions } from "../../auth/[...nextauth]";
 const cloudinary = require("cloudinary").v2;
 
 cloudinary.config({
